feat(turbopack): attach type metadata for static metadata item
